+++ pallets/manta-pay/src/test/payment.rs
@@ -160,9 +160,7 @@ where
         Some(id) => id,
         None => rng.gen(),
     };
-    // FIXME: get rid of the division after parity fixes the pallet-asset bug
-    let double_balance: u128 = rng.gen();
-    let total_free_balance = AssetValue(double_balance / 2);
+    let total_free_balance = AssetValue(rng.gen());
     let balances = value_distribution(count, total_free_balance, rng);
     initialize_test(asset_id, total_free_balance + DEFAULT_ASSET_ED);
     let mut utxo_accumulator = UtxoAccumulator::new(UTXO_ACCUMULATOR_MODEL.clone());
@@ -232,9 +230,7 @@ where
         Some(id) => id,
         None => rng.gen(),
     };
-    // FIXME: This is a workaround due to the substrate asset bug
-    let double_balance: u128 = rng.gen();
-    let total_free_balance = AssetValue(double_balance / 2);
+    let total_free_balance = AssetValue(rng.gen());
     let balances = value_distribution(count, total_free_balance, rng);
     initialize_test(asset_id, total_free_balance + DEFAULT_ASSET_ED);
     let mut utxo_accumulator = UtxoAccumulator::new(UTXO_ACCUMULATOR_MODEL.clone());
@@ -314,10 +310,7 @@ fn to_private_should_work() {
     let mut rng = OsRng;
     new_test_ext().execute_with(|| {
         let asset_id = rng.gen();
-        // FIXME: get rid of divide by two after parity fix pallet-asset
-        // This is to work around the substrate bug
-        let double_supply: u128 = rng.gen();
-        let total_free_supply = AssetValue(double_supply / 2);
+        let total_free_supply = AssetValue(rng.gen());
         initialize_test(asset_id, total_free_supply + DEFAULT_ASSET_ED);
         mint_tokens(
             asset_id,
@@ -331,10 +324,7 @@ fn to_private_should_work() {
 fn native_asset_to_private_should_work() {
     let mut rng = OsRng;
     new_test_ext().execute_with(|| {
-        // FIXME: get rid of divide by two after parity fix pallet-asset
-        // This is to work around the substrate bug
-        let double_supply: u128 = rng.gen();
-        let total_free_supply = AssetValue(double_supply / 2);
+        let total_free_supply = AssetValue(rng.gen());
         initialize_test(NATIVE_ASSET_ID, total_free_supply + DEFAULT_ASSET_ED);
         mint_tokens(
             NATIVE_ASSET_ID,
@@ -350,9 +340,7 @@ fn overdrawn_mint_should_not_work() {
     let mut rng = OsRng;
     new_test_ext().execute_with(|| {
         let asset_id = rng.gen();
-        // FIXME: remove the division after parity fix the pallet-asset bug
-        let double_supply: u128 = rng.gen();
-        let total_supply = AssetValue(double_supply / 2);
+        let total_supply = AssetValue(rng.gen());
         initialize_test(asset_id, total_supply + DEFAULT_ASSET_ED);
         assert_noop!(
             MantaPayPallet::to_private(